requirements.txt
Scrapy==0.22.2
scrapy-mongodb==0.7.1
scrapy-mongodb
